+++ Ogma3/Data/Models/Tag.cs
@@ -33,7 +33,7 @@ namespace Ogma3.Data.Models
 
         [ForeignKey("NamespaceId")]
         [JsonIgnore]
-        public  Namespace? Namespace { get; set; } = null;
-        public int? NamespaceId { get; set; } = null;
+        public Namespace? Namespace { get; set; } = null;
+        public long? NamespaceId { get; set; } = null;
     }
 }
