+++ Ogma3/Pages/Stories/Create.cshtml.cs
@@ -82,7 +82,7 @@ namespace Ogma3.Pages.Stories
             public int Rating { get; set; }
 
             [Required] 
-            public List<int> Tags { get; set; }
+            public List<long> Tags { get; set; }
         }
 
         public async Task OnGetAsync()
